feat: preloaded cache
